+++ Editor/AvatarProcessor.cs
@@ -3,6 +3,11 @@
 #if VRC_SDK_VRCSDK3
 
 using KRT.VRCQuestTools.Components;
+using KRT.VRCQuestTools.I18n;
+using KRT.VRCQuestTools.Models;
+using KRT.VRCQuestTools.Utils;
+using System.Text.RegularExpressions;
+using UnityEditor;
 using UnityEngine;
 using VRC.SDKBase.Editor.BuildPipeline;
 
@@ -25,6 +30,18 @@ namespace KRT.VRCQuestTools
         /// <returns>true.</returns>
         public bool OnPreprocessAvatar(GameObject avatarGameObject)
         {
+            if (VRCSDKUtility.CountMissingComponentsInChildren(avatarGameObject, true) > 0)
+            {
+                var i18n = VRCQuestToolsSettings.I18nResource;
+                var originalName = Regex.Replace(avatarGameObject.name, @"\(Clone\)$", string.Empty);
+                var result = EditorUtility.DisplayDialog("VRCQuestTools", i18n.MissingRemoverOnBuildDialogMessage(originalName), i18n.YesLabel, i18n.NoLabel);
+                if (!result)
+                {
+                    return false;
+                }
+                VRCSDKUtility.RemoveMissingComponentsInChildren(avatarGameObject, true);
+            }
+
             var components = avatarGameObject.GetComponentsInChildren<VertexColorRemover>(true);
             foreach (var c in components)
             {