+++ Editor/I18n/I18nEnglish.cs
@@ -62,6 +62,7 @@ namespace KRT.VRCQuestTools.I18n
         internal override string NoMissingComponentsMessage(string objectName) => $"There are no \"missing\" components in {objectName}.";
         internal override string MissingRemoverConfirmationMessage(string objectName) => $"Remove \"missing\" components from {objectName}.";
         internal override string UnpackPrefabMessage => "This also executes \"Unpack Prefab\" operation.";
+        internal override string MissingRemoverOnBuildDialogMessage(string objectName) => $"{objectName} has \"missing\" components. Would you like to continue building the avatar?";
 
         // BlendShapes Copy
         internal override string SourceMeshLabel => "Source Mesh";